fix(symthaea): CI hardening — psych-bench harness, liquid-mamba gate, sync fmt check
CI improvements pushing toward 111+/114 GREEN:
- Remove liquid-mamba from CI feature lists (missing candle deps in standalone)
- Add security benchmarks to harness completeness test (10 missing → 0)
- Add cargo fmt --check to sync-to-standalone.sh pre-push verification
- Scope CI-safe clippy/test to -p symthaea (avoid workspace member issues)
- Add school_learning --test-threads=1 to prevent timeout on free runners
- Increase test-feature-matrix timeout to 420 minutes
- Broca/spore clippy fixes (formatting, doc comments, gating)
- Psych-bench report/baselines/difficulty updates for new benchmark domains
- Therapeutic integration, substrate simulation, and test updates

+++ symthaea/crates/symthaea-psych-bench/src/benchmarks/mathematics/constraint_puzzles.rs
@@ -42,7 +42,7 @@ fn count_queens_violations(queens: &[usize]) -> u32 {
             // Same diagonal
             let row_diff = (queens[i] as isize - queens[j] as isize).unsigned_abs();
             let col_diff = j - i;
-            if row_diff == col_diff as usize {
+            if row_diff == col_diff {
                 violations += 1;
             }
         }
